requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
xgboost>=2.0
pyyaml>=6.0
nfl_data_py>=0.3.2
pyarrow>=16.0
joblib>=1.4
streamlit>=1.37
